+++ src/subservient_sensors.rs
@@ -1,0 +1,6 @@
+use bevy::prelude::*;
+
+#[derive(Component)]
+pub struct TnuaSubservientSensor {
+    pub owner_entity: Entity,
+}